requirements.txt
fastapi==0.115.12
gunicorn==23.0.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# --- Data Handling ---
pydantic==2.11.3
sqlmodel==0.0.25
Pillow>=10.0.0
joblib>=1.3.0
motor==3.4.0
pymongo==4.6.0
pydantic-settings==2.1.0
bcrypt==4.0.1
passlib[bcrypt]

numpy==1.24.3
opencv-python==4.8.1.78
minio==7.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
httpx==0.25.0
aiofiles==23.2.1
python-json-logger==2.0.7